pyproject.toml
[tool.poetry]
name = "ticket sales bot"
version = "0.1.0"
description = ""
authors = ["ilybuckskin <99vengeance99@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
aiogram = "^3.17.0"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.38"}
asyncpg = "^0.30.0"
alembic = "^1.14.1"
pydantic-settings = "^2.7.1"
qrcode = {extras = ["pil"], version = "^8.1"}
reportlab = "^4.4.0"
pyzbar = "^0.1.9"
pillow = "^11.2.1"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.4"
pytest-asyncio = "^0.25.3"
pytest-mock = "^3.14.0"
ruff = "^0.9.10"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.ruff]
line-length = 88
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "W", "I"]
ignore = []

[tool.ruff.format]
quote-style = "double"


[tool.pytest.ini_options]
pythonpath = ["."]
asyncio_mode = "auto"